requirements.txt
assemblyai>=0.33.0
python-dotenv
pyaudio
